worker_service/requirements-second.txt
aio-pika==9.5.5
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiormq==6.8.1
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.0.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.115.12
frozenlist==1.7.0
h11==0.16.0
hiredis==3.2.1
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.5.0
packaging==25.0
pamqp==3.3.0
passlib==1.7.4
pillow==11.3.0
pluggy==1.6.0
prometheus_client==0.22.1
propcache==0.3.2
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.4.0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.4.0
requests==2.32.4
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
yarl==1.20.1
